add and use Scope in JS symbolication
