+++ crates/symbolicator-service/tests/integration/sourcemap.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use symbolicator_service::types::JsFrame;
+use symbolicator_service::types::{JsFrame, Scope};
 use symbolicator_service::{
     services::symbolication::SymbolicateJsStacktraces, types::JsStacktrace,
 };
@@ -32,11 +32,12 @@ fn make_js_request(
     let stacktraces = vec![JsStacktrace { frames }];
 
     SymbolicateJsStacktraces {
+        scope: Scope::Global,
         source: Arc::new(source),
         stacktraces,
         modules: vec![],
         dist: dist.into(),
-        allow_scraping: true,
+        allow_scraping: false,
     }
 }
 
@@ -45,9 +46,8 @@ async fn test_sourcemap_expansion() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
     let (_srv, source) = symbolicator_test::sourcemap_server("01_sourcemap_expansion");
 
-    // NOTE: abs_path intentionally malformed to avoid fetching / applying source context
     let frames = r#"[{
-        "abs_path": "http//example.com/index.html",
+        "abs_path": "http://example.com/index.html",
         "filename": "index.html",
         "lineno": 6,
         "colno": 7,
@@ -83,10 +83,9 @@ async fn test_sourcemap_source_expansion() {
     let (symbolication, _cache_dir) = setup_service(|_| ());
     let (_srv, source) = symbolicator_test::sourcemap_server("02_sourcemap_source_expansion");
 
-    // NOTE: abs_path intentionally malformed to avoid fetching / applying source context
     let frames = r#"[{
         "function": "function: \"HTMLDocument.<anonymous>\"",
-        "abs_path": "http//example.com/index.html",
+        "abs_path": "http://example.com/index.html",
         "filename": "index.html",
         "lineno": 283,
         "colno": 17,
@@ -110,10 +109,9 @@ async fn test_sourcemap_embedded_source_expansion() {
     let (_srv, source) =
         symbolicator_test::sourcemap_server("03_sourcemap_embedded_source_expansion");
 
-    // NOTE: abs_path intentionally malformed to avoid fetching / applying source context
     let frames = r#"[{
         "function": "function: \"HTMLDocument.<anonymous>\"",
-        "abs_path": "http//example.com/index.html",
+        "abs_path": "http://example.com/index.html",
         "filename": "index.html",
         "lineno": 283,
         "colno": 17,
